simple convolution init

+++ FourierTransforms/SimpleConvolutionExample.py
@@ -1,13 +1,12 @@
 import numpy as np
 import matplotlib.image as mpimg
 import matplotlib.pyplot as plt
-from PIL import Image
 
 folder = "C:\\Users\\andre\\OneDrive - Politecnico di Milano\\Documenti\\PythonProjects\\Lectures\\FourierTransforms\\"
 filename = 'nyc.jpg'
 
 
-im = np.float32(mpimg.imread(folder+filename))
+im = mpimg.imread(folder+filename)
 plt.figure()
 plt.imshow(im)
 plt.title('original image')
